feat: move entity and notification to shared domain at core
