+++ src/core/video/application/use_cases/create_video_without_media.py
@@ -2,7 +2,7 @@ import uuid
 from dataclasses import dataclass
 from decimal import Decimal
 
-from src.core._shared.notification import Notification
+from src.core._shared.domain.notification import Notification
 from src.core.cast_member.domain.cast_member_repository import CastMemberRepository
 from src.core.category.domain.category_repository import CategoryRepository
 from src.core.genre.domain.genre_repository import GenreRepository
@@ -95,7 +95,6 @@ class CreateVideoWithoutMedia:
                 description=input.description,
                 launch_year=input.launch_year,
                 duration=input.duration,
-                published=input.published,
                 rating=input.rating,
                 categories=input.categories,
                 genres=input.genres,
